// apps/web/src/api-client/model/adminRideDetailStopsItem.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Waymate API
 * Carpooling backend powering the Waymate web app.
 * OpenAPI spec version: 0.1.0
 */
import type { RideStopId } from "./rideStopId";
import type { AdminRideDetailStopsItemCountryCode } from "./adminRideDetailStopsItemCountryCode";

export type AdminRideDetailStopsItem = {
    id: RideStopId;
    /**
     * @minimum -9007199254740991
     * @maximum 9007199254740991
     */
    stopOrder: number;
    address: string;
    city: string;
    /** @nullable */
    countryCode: AdminRideDetailStopsItemCountryCode;
    /** @nullable */
    plannedArrivalAt: string | null;
    /** @nullable */
    plannedDepartureAt: string | null;
};